// src/validate-technical-feasibility/responses/get-aba-data-response.interface.ts
import { GetABADataConstants } from '../constants/get-aba-data.constants';

export interface IGetABADataResponse {
  abadslamportid: number;
  abancc: string;
  abaclienttype: string;
  abaorderdate: string;
  abaad: string;
  abaparad: string;
  abaslot: number;
  abaport: number;
  abarack: number;
  abaposition: number;
  abavci: number;
  abacontractid: number;
  status: GetABADataConstants;
}